fix(diagnostics): restore dock auto-open after the lazy-load split
- Move the new-error watcher from DiagnosticsDock into an always-mounted useDiagnosticsAutoOpen hook in AppLayout
- The dock is lazy-mounted behind a first-open gate, so a watcher inside it could never see the first error and auto-open went dead
- Caught by the diagnostics auto-open, IPC error propagation, and error-retry E2E suites
